Ultrasonic Driver

+++ Firefighter-Robot_Code/HAL/inc/Ultrasonic.h
@@ -18,7 +18,7 @@
 
 
 void Ultrasonic_init(void);
-int Ultrasnic_Read(void);
+uint8_t Ultrasnic_Read(void);
